+++ src/Controller/SMTPTLS_ReportsController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
@@ -13,6 +14,8 @@ use Symfony\Contracts\Translation\TranslatorInterface;
 use App\Entity\Users;
 use App\Entity\SMTPTLS_Reports;
 use App\Entity\Domains;
+
+use App\Form\DeleteFormType;
 
 use App\Repository\SMTPTLS_ReportsRepository;
 
@@ -47,7 +50,7 @@ class SMTPTLS_ReportsController extends AbstractController
         if(isset($_GET["perpage"]) && $_GET["perpage"] > 0) {
             $pages['perpage'] = intval($_GET["perpage"]);
         } else {
-            $pages['perpage'] = 17;
+            $pages['perpage'] = 10;
         }
 
         $repository = $this->em->getRepository(Domains::class);
@@ -60,14 +63,19 @@ class SMTPTLS_ReportsController extends AbstractController
         } else {
             $reports = $repository->findBy(array('domain' => $domains), array('id' => 'DESC'), $pages["perpage"], ($pages["page"] - 1) * $pages["perpage"]);
         }
-        $totalreports = $repository->getTotalRows($reports);
+        $pages["totalitems"] = $repository->getTotalRows($reports);
+        $pages["start"] = $pages["totalitems"] - (($pages["page"] - 1) * $pages["perpage"]); 
+        $pages["end"] = $pages["totalitems"] - (($pages["page"] - 1) * $pages["perpage"]) - $pages['perpage'] + 1;
+        if ($pages["end"] < 0) {
+            $pages["end"] = 1;
+        }
 
-        if(count($reports) == 0 && $totalreports != 0 && $pages["page"] != 1) {
+        if(count($reports) == 0 && $pages["totalitems"] != 0 && $pages["page"] != 1) {
             return $this->redirectToRoute('app_smtptls_reports');
         }
 
-        $pages["total"] = ceil($totalreports / $pages['perpage']);
-        if($totalreports / $pages['perpage'] > $pages["page"]) {
+        $pages["total"] = ceil($pages["totalitems"] / $pages['perpage']);
+        if($pages["totalitems"] / $pages['perpage'] > $pages["page"]) {
             $pages["next"] = true;
         }
         if($pages["page"] - 1 > 0) {
@@ -77,7 +85,15 @@ class SMTPTLS_ReportsController extends AbstractController
         return $this->render('smtptls_reports/index.html.twig', [
             'reports' => $reports,
             'pages' => $pages,
-            'menuactive' => 'reports',
+            'page' => array(
+                'menu' => array(
+                    'category' => 'reports',
+                    'item' => 'smtptls'
+                ),
+                'pretitle' => $this->translator->trans("Reports"),
+                'title' => $this->translator->trans("SMTP-TLS reports"),
+                'actions' => array(),
+            ),
             'breadcrumbs' => array(
                 array('name' => $this->translator->trans("Reports"), 'url' => $this->router->generate('app_reports')),
                 array('name' => $this->translator->trans("SMTPTLS"), 'url' => $this->router->generate('app_smtptls_reports'))
@@ -97,7 +113,7 @@ class SMTPTLS_ReportsController extends AbstractController
         $userRepository = $this->em->getRepository(Users::class);
 
         if(!$userRepository->denyAccessUnlessOwned($repository->getDomain($report), $this->getUser())) {
-            return $this->render('not_found.html.twig', []);
+            return $this->render('base/error.html.twig', ['page' => array('title'=> 'Not found'), 'message' => $exception->getMessage()]);
         }
 
         if(!in_array($this->getUser(), $report->getSeen()->getValues())) {
@@ -107,7 +123,15 @@ class SMTPTLS_ReportsController extends AbstractController
         }
 
         return $this->render('smtptls_reports/report.html.twig', [
-            'menuactive' => 'reports',
+            'page' => array(
+                'menu' => array(
+                    'category' => 'reports',
+                    'item' => 'smtptls'
+                ),
+                'pretitle' => $this->translator->trans("Reports"),
+                'title' => $this->translator->trans("SMTP-TLS report #".$report->getId()),
+                'actions' => array(),
+            ),
             'breadcrumbs' => array(
                 array('name' => $this->translator->trans("Reports"), 'url' => $this->router->generate('app_reports')),
                 array('name' => $this->translator->trans("SMTPTLS"), 'url' => $this->router->generate('app_smtptls_reports')),
@@ -118,7 +142,7 @@ class SMTPTLS_ReportsController extends AbstractController
     }
 
     #[Route('/reports/smtptls/delete/{report}', name: 'app_smtptls_reports_delete')]
-    public function delete(SMTPTLS_Reports $report): Response
+    public function delete(SMTPTLS_Reports $report, Request $request): Response
     {
         if (!$this->getUser() || !$this->isGranted('IS_AUTHENTICATED')) {
             return $this->redirectToRoute('app_login');
@@ -126,9 +150,41 @@ class SMTPTLS_ReportsController extends AbstractController
 
         $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
-        $this->em->remove($report);
-        $this->em->flush();
+        $form = $this->createForm(DeleteFormType::class);
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid())
+        {
+            $formdata = $form->getData();
+            if($formdata['item'] == $report->getId()) {
+                $this->addFlash('success', 'Report deleted');
 
-        return $this->redirectToRoute('app_smtptls_reports');
+                $this->em->remove($report);
+                $this->em->flush();
+                
+                return $this->redirectToRoute('app_smtptls_reports');
+            } else {
+                $this->addFlash('danger', 'The id you typed does not match the report id');
+            }
+        }
+
+        return $this->render('base/delete.html.twig', [
+            'page' => array(
+                'menu' => array(
+                    'category' => 'domains',
+                    'item' => 'edit'
+                ),
+                'pretitle' => $this->translator->trans("SMTP-TLS Reports"),
+                'title' => $this->translator->trans("Delete report")." ".$report->getId(),
+                'actions' => array(),
+            ),
+            'item' => $report->getId(),
+            'form' => $form,
+            'breadcrumbs' => array(
+                array('name' => $this->translator->trans("Reports"), 'url' => $this->router->generate('app_reports')),
+                array('name' => $this->translator->trans("SMTP-TLS"), 'url' => $this->router->generate('app_smtptls_reports')),
+                array('name' => $this->translator->trans("Report")." #".$report->getId(), 'url' => $this->router->generate('app_smtptls_reports')),
+                array('name' => $report->getId(), 'url' => $this->router->generate('app_smtptls_reports_delete', ['report' => $report->getId()]))
+            ),
+        ]);
     }
 }
